Improve default pin settings for FRDM-KL25Z.
Pins that don't exist on the MCU are set to UNCONNECTED.
Pins that are not dedicated to a particular use are set to INPUT_ANALOG
(we don't want them to be Hi-Z inputs if they are unconnected).
Pins that have special purposes such as RESET_b and SWD pins are set to
their special functions.